util/frontend: change transfer rate to double
Motivation:

Long division of bytes transferred by seconds
yields zero for slow transfers.

See:

https://rt.dcache.org/Ticket/Display.html?id=10209

Modification:

Change to floating point division.

Result:

Clients are happier.

Target: master
Request: 7.2
Request: 7.1
Request: 7.0
Request: 6.0
RT: #10209
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/13215/
Acked-by: Lea

+++ modules/dcache/src/main/java/diskCacheV111/util/TransferInfo.java
@@ -230,17 +230,17 @@ public class TransferInfo implements Comparable<TransferInfo>, InvalidatableItem
         return subject;
     }
 
-    public Long getTransferRate() {
+    public Double getTransferRate() {
         if (bytesTransferred == null) {
-            return 0L;
+            return 0.0;
         }
 
         if (transferTime == null) {
-            return 0L;
+            return 0.0;
         }
 
-        long secs = transferTime / 1000;
-        return secs > 0.0 ? BYTES.toMiB(bytesTransferred) / secs : 0L;
+        double secs = transferTime / 1000.0;
+        return secs > 0.0 ? BYTES.toMiB((double) bytesTransferred) / secs : 0.0;
     }
 
     public Long getTransferTime() {
